Remove alt clause from cond form
The alt clause could sometimes be syntactically equivalent to an arm
with a predicate:

```
(cond ((= 2 4) '(this cannot be reached))
      (not true))
-> true
```

In this case `not` would be interpreted as a predicate instead of part
of a function application.  As not is defined as a function (and
consequently not nil), the consequence of the arm `true` would be
returned, instead of the intended `nil`.

For this reason the alt part of the cond form has been removed.  A way
to emulate the old behaviour is by setting the last predicate as `true`:

```
(cond ((= 2 4) '(this cannot be reached))
      (true    (not true)))
-> nil
```